Rename file to `ledger_context`
