+++ nano/core_test/confirming_set.cpp
@@ -6,7 +6,7 @@
 #include <nano/node/make_store.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/ledger_set_confirmed.hpp>
-#include <nano/test_common/ledger.hpp>
+#include <nano/test_common/ledger_context.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
