+++ nano/core_test/ledger.cpp
@@ -13,7 +13,7 @@
 #include <nano/secure/ledger_set_any.hpp>
 #include <nano/secure/ledger_set_confirmed.hpp>
 #include <nano/store/rocksdb/rocksdb.hpp>
-#include <nano/test_common/ledger.hpp>
+#include <nano/test_common/ledger_context.hpp>
 #include <nano/test_common/make_store.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
